Add keybind for `L` to play/fast-forward a track
If you L on a track that's not playing, it will start playing.

If you L on your current track, it will fast-forward 30 seconds.

If you L when you're not playing anything, it will error because it
doesn't have the brains to pick the proper device to start playback on.

Also required another revisit to fucking axios and it's inconsistent
handling of the different HTTP verbs. Blegh.