Show how long ago a track was saved in Home view
Instead of @current.saved being a boolean, make it a date so we can keep
track of when tracks get saved. And by "date", I mean a string so that
the history is serialized properly; even though JSON.stringify can
normally serialize a Date() to a string, it is failing to do so in my
case.